+++ module/testing/testing-damage-functions.js
@@ -2,9 +2,7 @@ import { HeroSystem6eActor } from "../actor/actor.js";
 import { HeroSystem6eItem } from "../item/item.js";
 import {
     determineStrengthDamage,
-    determineExtraDiceDamage,
     convertFromDC,
-    addTerms,
     convertToDcFromItem,
 } from "../utility/damage.js";
 
@@ -113,47 +111,6 @@ export function registerDamageFunctionTests(quench) {
                 });
             });
 
-            describe("Extra Damage", function () {
-                it("Extra Dice- Zero", function () {
-                    const item = new HeroSystem6eItem({
-                        name: "Test",
-                        type: "attack",
-                        system: {
-                            extraDice: "zero",
-                        },
-                        parent: actor,
-                    });
-
-                    assert.equal(determineExtraDiceDamage(item), "");
-                });
-
-                it("Extra Dice - Pip", function () {
-                    const item = new HeroSystem6eItem({
-                        name: "Test",
-                        type: "attack",
-                        system: {
-                            extraDice: "pip",
-                        },
-                        parent: actor,
-                    });
-
-                    assert.equal(determineExtraDiceDamage(item), "+1");
-                });
-
-                it("Extra Dice - Half", function () {
-                    const item = new HeroSystem6eItem({
-                        name: "Test",
-                        type: "attack",
-                        system: {
-                            extraDice: "half",
-                        },
-                        parent: actor,
-                    });
-
-                    assert.equal(determineExtraDiceDamage(item), "+1d3");
-                });
-            });
-
             describe("convertFromDC", function () {
                 describe("invalid inputs", function () {
                     const item = new HeroSystem6eItem({
@@ -185,7 +142,7 @@ export function registerDamageFunctionTests(quench) {
                     });
 
                     it("2", function () {
-                        assert.equal(convertFromDC(killingItem, 2), "1d3");
+                        assert.equal(convertFromDC(killingItem, 2), "½d6");
                     });
 
                     it("3", function () {
@@ -193,14 +150,11 @@ export function registerDamageFunctionTests(quench) {
                     });
 
                     it("4", function () {
-                        assert.equal(convertFromDC(killingItem, 4), "1d6 + 1");
+                        assert.equal(convertFromDC(killingItem, 4), "1d6+1");
                     });
 
                     it("5", function () {
-                        assert.equal(
-                            convertFromDC(killingItem, 5),
-                            "1d6 + 1d3",
-                        );
+                        assert.equal(convertFromDC(killingItem, 5), "1½d6");
                     });
 
                     it("6", function () {
@@ -208,7 +162,47 @@ export function registerDamageFunctionTests(quench) {
                     });
 
                     it("7", function () {
-                        assert.equal(convertFromDC(killingItem, 7), "2d6 + 1");
+                        assert.equal(convertFromDC(killingItem, 7), "2d6+1");
+                    });
+                });
+
+                describe("killing attacks with 1d6-1 rather than 1/2d6", function () {
+                    const killingItem = new HeroSystem6eItem({
+                        name: "Test",
+                        type: "attack",
+                        system: {
+                            killing: true,
+                            extraDice: "one-pip",
+                        },
+                        parent: actor,
+                    });
+
+                    it("1", function () {
+                        assert.equal(convertFromDC(killingItem, 1), "1");
+                    });
+
+                    it("2", function () {
+                        assert.equal(convertFromDC(killingItem, 2), "1d6-1");
+                    });
+
+                    it("3", function () {
+                        assert.equal(convertFromDC(killingItem, 3), "1d6");
+                    });
+
+                    it("4", function () {
+                        assert.equal(convertFromDC(killingItem, 4), "1d6+1");
+                    });
+
+                    it("5", function () {
+                        assert.equal(convertFromDC(killingItem, 5), "2d6-1");
+                    });
+
+                    it("6", function () {
+                        assert.equal(convertFromDC(killingItem, 6), "2d6");
+                    });
+
+                    it("7", function () {
+                        assert.equal(convertFromDC(killingItem, 7), "2d6+1");
                     });
                 });
 
@@ -230,21 +224,21 @@ export function registerDamageFunctionTests(quench) {
                     it("1.2", function () {
                         assert.equal(
                             convertFromDC(nonKillingItem, 1.2),
-                            "1d6 + 1",
+                            "1d6+1",
                         );
                     });
 
                     it("1.5", function () {
                         assert.equal(
                             convertFromDC(nonKillingItem, 1.5),
-                            "1d6 + 1d3",
+                            "1½d6",
                         );
                     });
 
                     it("13.2", function () {
                         assert.equal(
                             convertFromDC(nonKillingItem, 13.2),
-                            "13d6 + 1",
+                            "13d6+1",
                         );
                     });
 
@@ -255,43 +249,9 @@ export function registerDamageFunctionTests(quench) {
                     it("1234567890.2", function () {
                         assert.equal(
                             convertFromDC(nonKillingItem, 1234567890.2),
-                            "1234567890d6 + 1",
+                            "1234567890d6+1",
                         );
                     });
-                });
-            });
-
-            describe("Add Terms", function () {
-                it("'' ''", function () {
-                    assert.equal(addTerms("", ""), "");
-                });
-
-                it("null null", function () {
-                    assert.equal(addTerms(null, null), "");
-                });
-
-                it("1d6 null", function () {
-                    assert.equal(addTerms("1d6", null), "1d6");
-                });
-
-                it("null 1d6", function () {
-                    assert.equal(addTerms(null, "1d6"), "1d6");
-                });
-
-                it("1d6 1d6", function () {
-                    assert.equal(addTerms("1d6", "1d6"), "1d6 + 1d6");
-                });
-
-                it("1d6 ''", function () {
-                    assert.equal(addTerms("1d6", ""), "1d6");
-                });
-
-                it("'' 1d6", function () {
-                    assert.equal(addTerms("", "1d6"), "1d6");
-                });
-
-                it("1d6 1d6", function () {
-                    assert.equal(addTerms("1d6", "1d6"), "1d6 + 1d6");
                 });
             });